zensar_tasks/requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-WTF==1.0.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
PyJWT==2.3.0
Werkzeug==2.0.3
WTForms==3.0.1